requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
streamlit==1.45.1
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.65
langchain-experimental==0.3.4
langchain-ollama==0.3.3
langchain-text-splitters==0.3.7
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
python-multipart==0.0.20
watchdog==6.0.0
pillow==11.0.0
python-dotenv==1.0.1
pydantic==2.10.4
aiofiles==24.1.0
pandas==2.2.3
plotly==5.24.1
requests==2.32.3
pytest==8.3.4
pytest-asyncio==0.25.0

